data-producers/requirements.txt
kafka-python==2.0.2
pyyaml==6.0
python-dateutil==2.8.2